Create Plane_seating_code
We found this code to be easier to work with and created a seating arrangement with user input with single digits to allow for customers to choose the seat and to check price or the check price before seating.  Hopefully this will enable the purchase of more than one seat and make things more ethical for all.